pyproject.toml
[project]
name = "novel_writer"
version = "0.1.0"
description = "AI-powered novel idea generation tool"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "python-slugify>=8.0.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/novel_writer"] 